400 D题题解
By dark_n8,
Hint : Try to model the situation as a graph with rooms as edges and
switches as nodes.

Editorial : All rooms are represented as edges. Mark the edges as 1 if
the room is open else mark the edge as closed. The answer will be "YES"
if you can color the graph in such a manner that the edges having value
0 have both nodes under different color (if the door is locked then one
of the switches should be selected) and the edges having 1 have both
nodes under same color (if the door is unlocked you should either select
both switches or neither of them). For checking the same start a bfs
from switch 1 and toggle it and proceed, if you are able to color all of
the switches then the answer is "YES" else it is not possible.