406 C题
/*by PrinceOfPersia*/
For each state of monster (2n possible states, the position and whose
turn it is) we will determine if it will be won, lost, or stuck in loop
(the player whose turn it is, will win, lose, or the game will never end
if this state happens).

For this purpose, first each state with monster on 1 is lost. Then if we
consider a graph that its vertices are our 2n states, we will
recursively determine the answer for each vertex. If a state has an edge
to a lost state, it's won, and if all its edges are to won states, it's
lost. The vertices that are neither lost or won at the end have answer
"loop".

You can implement this using a simple memoize.