308 E题
by：Wild_Hamster
We can see, that we can reach maximal answer, when brackets will be
between two signs  * , or between one sign  *  and the end of
expression. For convenience we will add in the begin of expression 1 * ,
and in the end of expression  * 1.
After that we can iterate all possible pairs of signs  *  and count the
expression with putting brackets between two signs  *  for each pair.
We can use two variables x and y to count the value of expression, in
the begin x = 0, y = firstnumber, where firstnumber is first digit of
expression, then if next sign is  + , then x = y, y = nextnumber, and if
next sign is  * , then x = x, y = y * nextnumber. The value of
expression will be x + y, we can create function, like that, to count
expressions inside and outside the brackets.