413 C题题解
There are three ways — Arkady builds one fountain for coins and other
for diamonds, Arkady builds both fountains for coins and Arkady builds
both fountains for diamonds. In the end we need to choose a way with
maximum total beauty.

The first way is simple — we need to choose one fountain of each type
with maximum beauty on which is enough coins and diamonds which Arkady
has. To make it we can easily iterate though all given fountains.

The other two ways are treated similarly to each other. Consider the way
when Arkady builds both fountains for coins.

Let's write out all the fountains for a coins in a separate array. We
will store them as pairs — cost and beauty. Then sort the array in
ascending order of cost. Let Arkady will build fountains with cost p1
and p2 coins and p1 ≤ p2 and p1 + p2 ≤ c. Additionally we need an array
maxB, where maxBi equals to maximum fountain beauty on the prefix of
sorted fountains for coins which ends in position i.

The option when p1 equals to p2 should be considered separately. It can
be done in one iteration through the fountains — from fountains with
same cost we need to choose two with maximum beauty and update the
answer with the resulting value.

It is only left case when p1 < p2. Let's brute the fountain with cost
equals to p2. After that, Arcady will have c - p2 coins. After that with
help of binary search we need to find the fountains prefix which Arkay
can build for the remaining coins. With help of array maxB we can
determine the maximum beauty of fountain which Arkady can build.

Similarly, the third way is solved when both fountains will be built for
diamonds.