409 C题题解
First, let's deal with the infinite case. If the supply of power is at
least as big as the sum of demands, we can keep all devices alive
indefinitely.

Otherwise, let's binary search for the result. We can do binary search
since if we can keep all devices alive for E seconds, we can keep it
alive for any time less than E seconds.

Since all usage/charging is continuous, we can think about it as
"splitting" up the charge amount.

For each device and a fixed time T, we can compute the rate that we need
to charge it so that the device always has positive power. Well, it uses
ai power per second, so it uses a total of T * ai power. It currently
has bi power, so we need Xi = max(0, T * ai - bi) units of power. This
means we need Xi / T units of power per second to this device.

So, we just need to check that . If so, then it is possible to keep the
devices alive for T seconds, and we can use this to continue our binary
search.

Since we are binary searching on doubles, it is useful to just do it for
a fixed number of steps. We can approximate the max answer is somewhere
around 1014 and we need a precision of 10 - 4, so we need approximately
iterations.