414 C题题解
Consider the following greedy strategy. When it's Oleg's turn, he will
replace the frontmost question mark with his smallest letter. When it's
Igor's turn, he will replace the frontmost question mark with his
largest letter. At first glance, you might think that this works.
However, there's another case that we haven't considered.
Suppose Oleg has the letters {x, y, z} and Igor has the letters
{a, b, c}. According to our previous strategy, Oleg will place x as the
first letter. However, that's not optimal. He can place his letters at
the back and force Igor to place the first letter. The reason is because
the largest letter of Igor is not larger than the smallest letter of
Oleg. Thus, it is beneficial for Oleg to place his letters at the back
and force Igor to place his letters in front.
So, what exactly will the final string look like? We'll look at the
moves one by one. If at some point Oleg's smallest letter is still
strictly smaller than Igor's largest letter, then both player must put
their smallest (largest if it's Igor) letter as the frontmost letter.
Why? Suppose not, then on the next turn the other player will occupy
that spot with their best (smallest if Oleg, largest if Igor) letter,
and the resulting string will be worse for the current player. This
proves that greedy is correct in this case.
Now, what if Oleg's smallest letter is not smaller than Igor's largest
letter. In this case, both players will want to force the other player
to place their own letter at the beginning of the string. It can be
proven that in this case, each person will place their current worst
(largest if Oleg, smallest if Igor) letter at the back of the string in
the optimal strategy. Thus, we can calculate the final string starting
from this point and after that reverse this part and combine it with the
first part of the string where both players greedily place their best
letters in the beginning.