313 D题
Let us note that "equivalence" described in the statements is actually
equivalence relation, it is reflexively, simmetrically and transitive.
It is meant that set of all string is splits to equivalence classes.
Let's find lexicographic minimal strings what is equivalent to first and
to second given string. And then check if its are equals.
It is remain find the lexicographic minimal strings what is equivalent
to given. For instance we can do it such a way: